added support for 'privacy' => true
In our not-for-profit organization we want to be transparent on the amount of time spent on a per task basis. But for privacy reason there must be no easy way to generate a report on a per person basis.

The attached patch adds support for privacy mode by removing the work-column if the assignee of a bug is used in the projection or condition of a query.

To enabled privacy mode, add

$wgBugzillaReports = array(
  ...
  'privacy'     => "true"
);

to LocalSettings.php